% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{Area_Under_Curve}
\alias{Area_Under_Curve}
\title{Calculate the Area Under the Curve}
\usage{
Area_Under_Curve(x, y, method = c("trapezoid", "step", "spline"),
  na.rm = FALSE)
}
\arguments{
\item{x}{the x-points of the curve}

\item{y}{the y-points of the curve}

\item{method}{can be "trapezoid" (default), "step" or "spline"}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds}
}
\value{
Area Under the Curve (AUC)
}
\description{
Calculate the area under the curve.
}
\examples{
x <- seq(0, pi, length.out = 200)
plot(x = x, y = sin(x), type = "l")
Area_Under_Curve(x = x, y = sin(x), method = "trapezoid", na.rm = TRUE)
}

