% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Classification.R
\name{FBeta_Score}
\alias{FBeta_Score}
\title{F-Beta Score}
\usage{
FBeta_Score(y_true, y_pred, positive = NULL, beta = 1)
}
\arguments{
\item{y_true}{Ground truth (correct) labels vector}

\item{y_pred}{Predicted labels vector, as returned by a classifier}

\item{positive}{An optional character string for the factor level that
corresponds to a "positive" result}

\item{beta}{Weight of precision in harmonic mean}
}
\value{
F-Beta Score
}
\description{
Compute the F-Beta Score
}

