% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLmetrics.R
\docType{package}
\name{MLmetrics}
\alias{MLmetrics}
\alias{MLmetrics-package}
\title{MLmetrics: Machine Learning Evaluation Metrics}
\description{
A collection of evaluation metrics, including loss, score and utility functions, 
that measure regression and classification performance.
}

