% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Regression.R
\name{MSE}
\alias{MSE}
\title{Mean Square Error Loss}
\usage{
MSE(y_true, y_pred)
}
\arguments{
\item{y_true}{Ground truth (correct) target values vector}

\item{y_pred}{Estimated target values vector}
}
\value{
Mean Square Error Loss
}
\description{
Compute the mean squared error regression loss.
}

