% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classification.R
\name{Precision_micro}
\alias{Precision_micro}
\title{Precision (micro averaged)}
\usage{
Precision_micro(y_true, y_pred, labels = NULL)
}
\arguments{
\item{y_true}{Ground truth (correct) labels vector}

\item{y_pred}{Predicted labels vector, as returned by a classifier}

\item{labels}{An optional vector containing the list of the existent
(unique) labels.}
}
\value{
Precision (micro averaged)
}
\description{
Compute the precision score of multi-class problem using the "micro" average.
details: https://sebastianraschka.com/faq/docs/multiclass-metric.html
}
\examples{
labels <- c("Q1","Q2","Q3","Q4")
truth <- sample(labels, 10, replace = TRUE)
pred <- sample(labels, 10, replace = TRUE)
Precision_micro(y_pred = pred, y_true = truth, labels)
}
