% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Regression.R
\name{R2_score}
\alias{R2_score}
\title{R-Squared (Coefficient of Determination) Regression Score}
\usage{
R2_score(y_true, y_pred)
}
\arguments{
\item{y_true}{Ground truth (correct) target values vector}

\item{y_pred}{Estimated target values vector}
}
\value{
R^2 score
}
\description{
Compute the R-Squared (Coefficient of Determination) Regression Score.
}

