% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classification.R
\name{Recall_macro_weighted}
\alias{Recall_macro_weighted}
\title{Recall (macro weighted averaged)}
\usage{
Recall_macro_weighted(y_true, y_pred, labels = NULL)
}
\arguments{
\item{y_true}{Ground truth (correct) labels vector}

\item{y_pred}{Predicted labels vector, as returned by a classifier}

\item{labels}{An optional vector containing the list of the existent
(unique) labels.}
}
\value{
Recall (macro weighted averaged)
}
\description{
Compute the recall score of multi-class problem using the "macro" formula but weighted averaged (similar to weka).
details: https://sebastianraschka.com/faq/docs/multiclass-metric.html
}
\examples{
labels <- c("Q1","Q2","Q3","Q4")
truth <- sample(labels, 10, replace = TRUE)
pred <- sample(labels, 10, replace = TRUE)
Recall_macro_weighted(y_pred = pred, y_true = truth, labels)
}
