% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Regression.R
\name{RMSE}
\alias{RMSE}
\title{Root Mean Square Error Loss}
\usage{
RMSE(y_true, y_pred)
}
\arguments{
\item{y_true}{Ground truth (correct) target values vector}

\item{y_pred}{Estimated target values vector}
}
\value{
Root Mean Square Error Loss
}
\description{
Compute the root mean squared error regression loss.
}
\examples{
data(cars)
reg <- lm(log(dist)~log(speed), data=cars)
RMSE(y_true=log(cars$dist), y_pred=reg$fitted.values)
}

