% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Regression.R
\name{RRSE}
\alias{RRSE}
\title{Root Relative Squared Error Loss}
\usage{
RRSE(y_true, y_pred)
}
\arguments{
\item{y_true}{Ground truth (correct) target values vector}

\item{y_pred}{Estimated target values vector}
}
\value{
Root Relative Squared Error Loss
}
\description{
Compute the root relative squared error regression loss.
}
\examples{
data(cars)
reg <- lm(log(dist)~log(speed), data=cars)
RRSE(y_true=log(cars$dist), y_pred=reg$fitted.values)
}

