% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Regression.R
\name{MAE}
\alias{MAE}
\title{Mean Absolute Error Loss}
\usage{
MAE(y_true, y_pred)
}
\arguments{
\item{y_true}{Ground truth (correct) target values vector}

\item{y_pred}{Estimated target values vector}
}
\value{
Mean Absolute Error Loss
}
\description{
Compute the mean absolute error regression loss.
}
\examples{
data(cars)
reg <- lm(log(dist)~log(speed), data=cars)
MAE(y_true=log(cars$dist), y_pred=reg$fitted.values)
}

