% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Regression.R
\name{MedianAE}
\alias{MedianAE}
\title{Median Absolute Error Loss}
\usage{
MedianAE(y_true, y_pred)
}
\arguments{
\item{y_true}{Ground truth (correct) target values vector}

\item{y_pred}{Estimated target values vector}
}
\value{
Median Absolute Error Loss
}
\description{
Compute the median absolute error regression loss.
}

