% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression.R
\name{RAE}
\alias{RAE}
\title{Relative Absolute Error Loss}
\usage{
RAE(y_pred, y_true)
}
\arguments{
\item{y_pred}{Estimated target values vector}

\item{y_true}{Ground truth (correct) target values vector}
}
\value{
Relative Absolute Error Loss
}
\description{
Compute the relative absolute error regression loss.
}
\examples{
data(cars)
reg <- lm(log(dist) ~ log(speed), data = cars)
RAE(y_pred = exp(reg$fitted.values), y_true = cars$dist)
}

